+++ .gitignore
@@ -9,4 +9,3 @@ build/
 mirrorverse.egg-info/
 *.pkl
 *.csv
-dvc.lock
